Work with nonroot invocation images
This updates the helm3 mixin for the upcoming change in Porter (v1)
where bundles will run as the nonroot user instead of root.
The kubeconfig file is now loaded using the same logic as kubectl,
using the KUBECONFIG environment variable, then $HOME/.kube/config.

It also switches the user that some commands run as during build.
Installing helm should run as root, but initializing the helm
repositories should run as the same user that the container runs as,
nonroot.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>